AI
Added some AI stuff.
Moved files around

RunnerBrain
An enemy that runs away from you
This is the main enemy of the game.
They can create a room and spawn minions

ChaserBrain
They chase you.
Nothing happens when they collider yet.

States
Made some states that might be useful

MapGen
Blocks are up at the start instead of down (this is how it is supposed to
be)

Motor
AI doesn't really use the motor. It's not easy to decouple the unity
pathfinding and movement (like just get the path as a series of points)
So it uses that as the movement.
Motor updates the speed over time like it does for the player

You now move faster if you look where you're going. Kind of op for the
enemies. They always look where they're going. There's a multiplier to
sclae how much it speeds you up by so theirs could just be 0.