Zigzag Diggin
When the runner runs out of minions to hide behind they try to dig to a
different area of the map.
They dig in a zigzag pattern to get there (instead of just digging
straight)
At the end of the tunnel, they wait for the player to get near before
bulding a new room with new minions.
The game repeats from this point

The runner starts in the bottom right of the map and moves to the 4
quadrants in clockwise order.
Breaks at the last quad (which is fine. I don't want it to continue
forever).

Tweaks
Blocks move faster
Digger bullet moves faster
With these, the runner should stand still less when digging

FollowTunnel
Does nothing. DigTunnel just does everything.
Wanted to dig, follow, dig, follow, etc. but it was easier to manage in 1
state with the zigzagging.

WaitForPlayer
Keeps time of how long it's waiting
Maybe in the future, it will go look for the player after 30s