Rules
Have a rule that halves damage when inside it

The rule follows the runner wizard around.

IRuleable
health implements this. Anything that has a health can be affected by
rules.
I wonder if I want 1 interface for all rules or differnt interfaces for
each rule?
Maybe health isn't the best place either. I just used it bc everything has
health.
But a rule that makes you faster would want Motor.
That does lean towards having an interface for each rule. I have no idea.
This is the first time I've used interfaces.

RadiusIndicator
Just draw a circle using a lineRenderer
Stolen from the internet.

(Rant) This was somehow way too hard to find. And the solution was on one of the
first pages I tried, but I made the mistake of scrolling down farther,
seeing solution 4 used shaders and thinking I could use shaders. What a
waste of time. And then when I used this solution and deleted the shader
stuff in frustration, it broke the materials in the particles (quitting
and reloading fixed it).