9 Quads, Gun Pickups
Using 9 quads bc it fits the arc of the game better.
Had to hack together some edge cases for the 9 quads instead of the 4
quads.

GameScript
Will be the script that drives the narrative of the game.
Loads the starting map, lets the player do a tutorial and then starts up
the AI.

DigTunnel
Stuck Timer. AI can get stuck on dead enemies and the non-solid chain
sometimes. If it takes longer than 7sec to get to the breach point, just
teleport it. Only way I really know how to fix it. I could despawn the
enemies, but I kinda like leaving them.

GunPickup
Can pick up guns by walking over the pickups.
Can drop your gun and spawn a pickup.
With this, I can make more guns and have a rule to drop them.