Minions with Rules, CantShootRule
Minions can spawn dragging around a rule
Doesn't work for ranged minions

GameScript
Made most of the game.
Rules spawn in the right order
Enemies spawn in specific numbers.
Some enemies carry around rules

Player
Can press G to respawn on command. For if you get stuck or fall out of the
map, etc.

CanShoot
When in the circle, you can't shoot. Should also affect the ranger
enemies.

Runner
Shouldn't get stuck on the last room. Was trying to get to the center
block's position (which is underground bc the center room is excavated at
the start)

Blocker Forcefield
Spawns closer to the center so should more reliably block the hole. By the
time it fails, the player knows what it is and what it should do so they
might not even notice they can walk around before they kill it