Room Spell and Bug Fixes
Cast a spell when you create a room
Make a sphere that quickly expands to the size of the room to make it look
like the runner cast a spell
Wait a few frames before making the room while this expands
Also wait a frame before making the rule for the room and spawning enemies

Bug Fixes
Did you know the runner was on the defaul layer this whole time? Because I
sure didn't. This whole time, it was just clippping through the forcefield
meant to keep the player out.
Now the forcefield layer doesn't collide with default so the runner walks
through it nice and smooth

Center of Room
Fixed it to use the runner position
Well, turns out the runner's y pos can sometimes be not 0.58 so I just set
it to 0.58 which is the height the runner is usually at (and most likely
to be at when near the center of the next room)